Update gas calculation for staking precompile
Please see comment in core/vm/contracts_write.go RequiredGas
